Sap(begin)

+++ ivan_purshega_web/src/main/java/dao/UserDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 import data.User;
 
 public interface UserDao {
-	    Long create(User user);
-	    User  read(Long id);
+	    int create(User user);
+	    User  read(int id);
 	    void update(User user);
 	    void delete(User user);
 	    List<User> findAll();
